+++ NEWS.md
@@ -1,3 +1,18 @@
+# DataFrames.jl v1.4 Release Notes
+
+## New functionalities
+
+* `permutedims` now supports a `strict` keyword argument that allows
+  for a more flexible handling of values stored in a column that will
+  become a new header
+  ([#3004](https://github.com/JuliaData/DataFrames.jl/issues/3004))
+* `unstack` now allows passing a function in `valuestransform` keyword argument;
+  this allows for a convenient creation of two dimensional pivot tables
+  ([#2998](https://github.com/JuliaData/DataFrames.jl/issues/2998))
+* Add special syntax for `eachindex`, `groupindices`, and `proprow`
+  to transformation mini-language
+  ([#3001](https://github.com/JuliaData/DataFrames.jl/pull/3001)).
+
 # DataFrames.jl v1.3.2 Patch Release Notes
 
 ## Bug fixes